Added performance measuring code
Can now perform K-fold cross validation on a subset of the data!
Now using a Random Forest Classifier, but maybe we want to use something else?
Using NDCG@38, as the Kaggle site specifies, over 25000 unique srch_id with 10-fold cross-validation (test_set is 2500 and training_set the remaining 22500 srch_ids) we obtain a score of just below 0.5 with the current code :)